Add files via upload
I'm excited to announce that I've completed the Beginner Level Figma Designs Task 02 of the YoungDev Intern Internship. This task involved creating a mobile app wireframe using Figma, and it has been an enriching experience!

Key takeaways from this task include:

🔹 Mastering the basics of mobile app wireframing
🔹 Enhancing my proficiency in Figma
🔹 Understanding user flow and interface design principles

A big thank you to the YoungDev team for this fantastic opportunity to grow and learn. Looking forward to the next challenge!